Fixed writing of results to CSV file to write as ANSI text. Fixed file extension used in results saving file dialog. Fixed Find command and added support for wrapping. Corrected use of int instead of size_t.

+++ AppCmds.hpp
@@ -119,7 +119,7 @@ protected:
 	//
 	// Internal methods.
 	//
-	void Connect(int nConnection, const CString& strLogin, const CString& strPassword);
+	void Connect(size_t nConnection, const CString& strLogin, const CString& strPassword);
 	bool LoadQuery(const CPath& strPath);
 	bool SaveQuery(const CPath& strPath);
 	void UpdateScriptsMenu();
